--- src/test/Node.css
@@ -1,14 +1,0 @@
-.visited {
-  border-radius: 20%;
-  box-shadow: inset 0 0 12px 5px rgba(36, 137, 81, 0);
-  animation: visit 0.3s ease forwards;
-}
-
-@keyframes visit {
-  0% {
-    box-shadow: inset 0 0 12px 5px rgba(36, 137, 81, 0);
-  }
-  100% {
-    box-shadow: inset 0 0 12px 5px rgba(36, 137, 81, 0.12);
-  }
-}
